% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inputvariable}
\alias{Inputvariable}
\title{Inputvariable}
\format{
An object of class \code{data.frame} with 17520 rows and 18 columns.
}
\usage{
Inputvariable
}
\description{
Just test dataframe(Description)
}
\examples{
head(Inputvariable)
}
\keyword{datasets}
