% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/result.R
\docType{data}
\name{result}
\alias{result}
\title{Information about model output}
\format{
A data frame with 1 rows and 31 variables:
\describe{
  \item{X}{sequence}
  \item{DATE}{date}
  \item{Vms}{The wind speed at measured height, m s-1}
  \item{Ta}{air temperature, degrees Celsius}
  \item{RH}{relative humidity, percent}
  \item{VPDhpa}{vapor pressure deficit, hPa}
  \item{SVWC30cm}{soil volumetric moisture at depth of 30 cm, percent}
  \item{Rn}{net radiation at the canopy surface, W m-2}
  \item{PPFD}{photosynthetic photon flux density, umol m-2 s-1}
  \item{Rainfall}{rainfall,mm}
  \item{Month}{the number of month}
  \item{Day}{the day of month}
  \item{year}{studied year}
  \item{time}{the time of day at 30 min scale, h}
  \item{DOY}{day of year}
  \item{Cof}{carbon dioxide concentration in the atmosphere, ppm}
  \item{G}{Soil heat flux, W m-2}
  \item{NEE}{Observed net ecosystem productivity at 30 min scale, mg CO2 m-2 s-1}
  \item{LE}{Observed latent heat at 30 min scale}
  \item{ObserveNEE30}{observed net ecosystem production, gC m-2 30 min-1}
  \item{OETS}{observed evapotranspiration, mm 30 min-1}
  \item{NEP30min}{net ecosystem production, gC m-2 30 min-1}
  \item{ETS}{evapotranspiration, mm 30 min-1}
  \item{GPP30min}{gross primary production, gC m-2 30 min-1}
  \item{Re30min}{ecosystem respiration, gC m-2 30 min-1}
}

A data frame with 1 rows and 31 variables:
\describe{
  \item{X}{sequence}
  \item{DATE}{date}
  \item{Vms}{The wind speed at measured height, m s-1}
  \item{Ta}{air temperature, degrees Celsius}
  \item{RH}{relative humidity, percent}
  \item{VPDhpa}{vapor pressure deficit, hPa}
  \item{SVWC30cm}{soil volumetric moisture at depth of 30 cm, percent}
  \item{Rn}{net radiation at the canopy surface, W m-2}
  \item{PPFD}{photosynthetic photon flux density, umol m-2 s-1}
  \item{Rainfall}{rainfall,mm}
  \item{Month}{the number of month}
  \item{Day}{the day of month}
  \item{year}{studied year}
  \item{time}{the time of day at 30 min scale, h}
  \item{DOY}{day of year}
  \item{Cof}{carbon dioxide concentration in the atmosphere, ppm}
  \item{G}{Soil heat flux, W m-2}
  \item{NEE}{Observed net ecosystem productivity at 30 min scale, mg CO2 m-2 s-1}
  \item{LE}{Observed latent heat at 30 min scale}
  \item{ObserveNEE30}{observed net ecosystem production, gC m-2 30 min-1}
  \item{OETS}{observed evapotranspiration, mm 30 min-1}
  \item{NEP30min}{net ecosystem production, gC m-2 30 min-1}
  \item{ETS}{evapotranspiration, mm 30 min-1}
  \item{GPP30min}{gross primary production, gC m-2 30 min-1}
  \item{Re30min}{ecosystem respiration, gC m-2 30 min-1}
}
}
\usage{
result

result
}
\description{
A dataset containing the list of output variables and their description.

A dataset containing the list of output variables and their description.
}
\keyword{datasets}
