% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inputvariable}
\alias{Inputvariable}
\title{Information about input variables}
\format{
A data frame with 17520 rows and 18 variables:
\describe{
  \item{DATE}{date}
  \item{Vms}{The wind speed at measured height, m s-1}
  \item{Ta}{air temperature, degrees Celsius}
  \item{RH}{relative humidity, \%}
  \item{VPDhpa}{vapor pressure deficit, hPa}
  \item{SVWC30cm}{soil volumetric moisture at depth of 30 cm, \%}
  \item{Rn}{net radiation at the canopy surface, W m-2}
  \item{PPFD}{photosynthetic photon flux density, μmol m-2 s-1}
  \item{Rainfall}{rainfall,mm}
  \item{Month}{the number of month}
  \item{Day}{the day of month}
  \item{year}{studied year}
  \item{time}{the time of day at 30 min scale, h}
  \item{DOY}{day of year}
  \item{Cof}{carbon dioxide concentration in the atmosphere, ppm}
  \item{G}{Soil heat flux, W m-2}
  \item{NEE}{Observed net ecosystem productivity at 30 min scale, mg CO2 m-2 s-1}
  \item{LE}{Observed latent heat at 30 min scale}
}
}
\usage{
Inputvariable
}
\description{
A dataset containing the list of input variables and their description.
}
\keyword{datasets}
